requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
Flask_Testing==0.8.1
langchain==0.0.220
openai==0.27.2
pinecone_client==2.2.1
PyPDF2==3.0.1
python-dotenv==1.0.0
reportlab==4.0.4
Requests==2.31.0
streamlit==1.22.0
Werkzeug==2.2.3
azure-core==1.26.4
azure-cosmos==4.3.1
ipinfo==4.4.2
python-jose==3.3.0
PyJWT==2.8.0